sensevoice_requirements.txt
# SenseVoice OpenAI-Compatible API Requirements

# Core dependencies from original requirements.txt
torch<=2.3
torchaudio
modelscope
huggingface
huggingface_hub
funasr>=1.1.3
numpy<=1.26.4

# API specific dependencies
fastapi>=0.111.1
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1
